support for async helpers with express 3.x
Normal handlebars helpers must return a value immediately, however it is
sometimes useful to be able to do some async stuff in a helper.
Handlebars does not support this natively, so we write a small wrapper
that allows for registrations of helpers which must call a callback when
finished. The final template is not rendered until these async helpers
finish as would be expected.