Create a new async per middleware
Need to try and figure out a reproducible test case, but I'm having issues using with Express 4, with view caching on, and editing the template. Getting the __async_ids__ sput out the first time after updating a template, trying to figure out why, but thought you might want to know. Somehow the `values` keys that come back to the @render_file/self.async.done callback don't match the ones in `res`. Will keep investigating, but again, this solves it for me. Thanks-