Fixing invalid check on undefined that can lead to bugs
The correct way to check if a variable is not defined, in Javascript is: ```typeof layout === 'undefined'```. Doing something like ```layout === undefined``` doesn't check if a variable is undefined, it leverages the fact that there's typically no variable called "undefined"and compares layout to essentially null. It works sometimes, almost by accident but it's essentially a bug waiting to happen.