Fix localsAsTemplateData when caching is enabled
When a template is rendered from the cache, the extra options object
(containing the data property) isn't being passed in to the render
function. Thus, any functionality that depends on this will fail.

The code has been refactored to create a single handlebars options
object for each (main) template render, which is then passed in all
scenarios (cache, template, layout template).

The code also cleans up the use of parent-scope variables for the render
context, and avoids a circular reference to the "_locals" object from
itself in favor of an instance-based boolean to determine whether the
locals should be passed as template data.

Tests have also been added to test for this scenario.

fixes https://github.com/donpark/hbs/issues/101